Create Functions
#Functions:
Functions are reusable blocks of code that perform a specific task. They can be called multiple times within a program, making code more efficient and easier to read.

#Syntax:
To create a function in JavaScript, use the keyword "function", followed by the function name and any parameters in parentheses. The code that the function executes is enclosed in curly braces {}.

#Calling a function:
To call a function, simply type the function name followed by parentheses (). If the function has parameters, include them inside the parentheses.

#Return statement:
Functions can also return a value using the "return" keyword. This allows the function to pass information back to the code that called it.

#Anonymous functions:
Anonymous functions are functions without a name. They are often used as callbacks, which are functions passed as arguments to other functions. Anonymous functions are defined using the "function" keyword without a name.

#Arrow functions:
Arrow functions are a shorter way to write functions in JavaScript. They use the => syntax instead of the "function" keyword, making them more concise and easier to read.